Manually added MeshSurfaceFilter to current score
Added MeshSurfaceFilter to current score through an if statement.
This is a non-domain specific filter, but it uses the same mesh
as the domain_filter. Therfore, I add the filter
openmc.MeshSurfaceFilter(domain_filter.mesh).
Next, I want to add this filter in the body of the SurfaceMXGS class
so that it is not "hard coded"