Re-wrote how MeshSurfaceFilter is added to current score
Instead of the unelegant if-statement in the loop that adds
non-domain specific filters, I added a filters property to
SurfaceMGXS to add the MeshSurfaceFilter. I still use the same
mesh as the domain-specific MeshFilter, and the way I refer to
it could probably be improved. It's a little bulky right now.