COMPATIBILITY OF MESHSURFACEFILTER WITH TRANSIENT CODE
MeshSurfaceFilter does not need to be use in conjunction with a MeshFilter.
So the MeshFilter was removed from the current tally.

The filters for the current tally were re-ordered to facilitate reshaping later on.

The reshape had to be done manually. When SurfaceFilter was used, the cell and
surface filters were separable, but this is not longer the case with MeshSurfaceFilter.
However, the code relies on the tally order: cell, energy, surface. Therefore,
the reshaping the tally into to the proper order had to be done manually with a nested for loop.