Added variables to settings.py
Added frequency variables to settings.py in the:
-attributes
-initialization
-properties
-setters
-xml subelements

I am still missing the from_xml functions.
I also need to improve the _create_frequency_subelement function.
Currently, it explicitly defines a mesh in the subelement. It would be better
if the code could just read a mesh id. Once the kinetics code is re-written to
accommodate this, I will make this change.

I also need to work on the code because it currently requires lower_left and width
be set for the frequency mesh. Ideally, it would accept both width and upper_right.
Once that change is made, I should revisit how the mesh is defined in settings.py