Working & tested version of frequency variables in settings.py
Modified from_xml() function for frequency variables.
Tested succesfully in test_settings.py.
It's not very succinct, but it works. In the future, I would like
to re-design the frequency xml entry to accept a mesh id instead of
an informal mesh defition. The bottleneck here is that the same mesh
is used for settings.xml and tallies.xml, which is not permitted due
to unique ids.
I also modified Sam's code to accept upper_right as well as width
for defining unity_mesh in solver.py.