Finished adding frequency variables to settings.py
1. frequency mesh. Unfortunately, it is still the "outdated" method of
fully definig a mesh in the <frequency> xml element.
Future work should allow user to define just the mesh ID.
However, this mesh is used in both settings and tallies, which is
currently not allowed (violates unique ID).
Also there might be an error in my code with mesh.shape_ if the dimensions
aren't specified.

2. group structure. Reads the energy groups needed for frequencies in CMFD

3. num delayed groups. Reads the number of delayed groups.

4. precursor frequency. Takes place within delayed group statement
because the number of delayed groups is needed. The presursor frequencies
are updated in the settings xml file as they are calculated throughout
the transient simulation.

5. flux frequency. The flux frequency is updated in the settings xml
file as it is calculated throughout the transient simulation.
Shouldn't I be able to have a flux frequency for each energy group?