Fixed indexing issues, still problems with delayed_nu_fission
I realized I have to 0-index on micro.delayed_nu_fission even though
the delayed groups are numbered 1-6 (0 is actually "prompt").
So I fixed that. I also realized that xs(i_grid, XS_DELAYED_NU_FISSION)
needs a "d-1" at the end of it to differentiate between the groups.

However, the problem of nu_fission != prompt+delayed persists.

I believe it lies with the delayed data, as it requires a different shape
than delayed and I've probably implemented it wrong.