Finished adding frequency modifications to particle.cpp
Put finishing touches on weight adjustment. This included adding
freq to particle.h so that its value would be assigned and used in
other scopes.

Tested on pin-td.py and saw that the frequencies matched the settings.py
output. Note that in many places, the final value of the frequency
is actually flux_frequency[freq_group]/velocity.

I still have not validated that macro_xs_.inverse_velocity behaves as
expected.