Cleaned up loops and if-statements. Fixed an issue with the size of delayed_nu_fission
In material.cpp I was resizing delayed_nu_fission for every nuclide
which is not appropriate for a material, where each nuclide has a
contriubtion to a fixed size vector.

I also eliminated many unnecessary if-statements and variable declarations.

I still haven't added in the change to the random number sequence in
physics_mg.cpp