Added last frequency modification to physics_mg
The last chunk of code modifies the random number sequence
so I waited until the very end to add it in.

I have verified that inverse_velocity works properly.
I also showed that at least one mgxs_library test can pass
when the new versions of the tests & mgxs.py & mdgxs.py are used
in conjunction with all the physics changes made here.

Unfortunately, at the moment the FTM seems to fail for a multi-group
problem run in Omega-mode with flux frequencies (fine if flux frequencies
are left out and only precursor frequencies are included). The error
is that the fission bank is full. Something to do with the sampling
scheme.