Fixed issue with i_nuclide
In the fortran code, i_nuclide=0 was a special case of not being assigned.
However, in the c++ code, due to 0-indexing, I needed the same case to
be changed to i_nuclide=-1.
Additionally, the snipet of code was re-arranged. In the fortran code,
a "while prob < cutoff" made sure that the loop would not execute if
the frequency was greater than the cutoff. However in the c++ code,
a static loop was used which only exited if prob exceeded cutoff.
The consequence was that i_nuclide was always assigned even if the
loop exited after the first iteration. I fixed that by adding a
condition to return i_nuclide as -1 if the frequency was larger
than the cutoff before the loop had a chance to start.