Fixed multi-group mode
The tally was erroneously acculumating the score, but updating the tally after
each accumulation, resulting in double accumulation. Instead, my edit
allows the score to be calculated by itself, update the tally, and then
calculate the new score.

In mdgxs.py, the reversal of the order only applies for energy, not
delayed groups. Therefore, it was wrongly reversing the order.