Updated the frequency mesh in settings.py, settings.cpp
Made setting the frequency mesh more in line with the modern OpenMC style - in keeping with the example set by entrop_mesh.

Since the same mesh can't be used in tallies.xml and settings.xml, I had to create a new mesh that is identical to the shape_mesh in solver.py.